LayoutTests/fast/css/CSSStyleDeclaration-setProperty-unset-important.html

<!DOCTYPE html>
<html>
<body>
<script src="../../resources/js-test-pre.js"></script>
<script>
description("Test that CSSStyleDeclaration.setProperty() can unset the 'important' flag");

var div = document.createElement("div");
evalAndLog("div.style.setProperty('color', 'red', 'important')");
shouldBeEqualToString("div.style.getPropertyValue('color')", "red");
shouldBeEqualToString("div.style.getPropertyPriority('color')", "important");

evalAndLog("div.style.setProperty('color', 'green', null)");
shouldBeEqualToString("div.style.getPropertyValue('color')", "green");
shouldBeEqualToString("div.style.getPropertyPriority('color')", "");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
