LayoutTests/fast/css/CSSStyleDeclaration-setProperty-unset-important-expected.txt

Test that CSSStyleDeclaration.setProperty() can unset the 'important' flag

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


div.style.setProperty('color', 'red', 'important')
PASS div.style.getPropertyValue('color') is "red"
PASS div.style.getPropertyPriority('color') is "important"
div.style.setProperty('color', 'green', null)
PASS div.style.getPropertyValue('color') is "green"
PASS div.style.getPropertyPriority('color') is ""
PASS successfullyParsed is true

TEST COMPLETE
